Flesh out CurrentClusters in CapacityDesires input.
We now have cpu, memory, and network utilization in the CurrentCapacity section of the CapacityDesires input. Furthermore the generate_scenarios() call which filters out instances now takes that shape into account.

This allows for callers of the plan api to (optionally) input the current cluster capacity and use that to inform instance selection. Models can furthermore use this to determine cluster size based on current utilziation.